requirements.txt
streamlit==1.55.0
openai==2.26.0
google-genai==1.66.0
anthropic==0.84.0
python-dotenv==1.2.2
pandas
inspect-ai>=0.3.188

